Add edit distance matching for lookalike URLs using top 500 domains
This is a follow up to crrev/1379195. It uses the .cc file generated in the previous CL to check if any top domains are within 1 edit distance of the navigated domain. The check is done as a linear search, as described in the design doc in the previous CL.

Bug: 913647
Change-Id: Ia78a079e786703678ef93c6f341138e15d074a6f
Reviewed-on: https://chromium-review.googlesource.com/c/1378973
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#617698}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5b516e05f9a3b3d0a4535c2a4ad15aa0e630be86