[signin] Export credentials to new profile during signin flow
This CL restores the pre-Dice behavior when a new profile is created
as part of the signin flow.

There are a few small differences:
1) when the token is removed from the profile to be exported to the new
profile, the account reconcilor also invalidates the cookie. As a
result the user will be signed out of the web in the source profile.

2) Device ID concerns. These issues were existing in the pre-Dice flow:
the token has the device ID of the source profile, and if the user
signs-in again in the source profile with the same account, then there
are now two tokens with the same device ID in two different profiles.
This somehow creates a link between these profiles:
- revoking a token revokes all the other tokens with the same device ID,
  and thus signing out in one profile would also sign out the other
  profile.
- there are privacy implications as Gaia could theoretically use this
  information to detect that two profiles are on the same computer: if
  profile 1 has tokens A and B, and profile 2 has tokens A and C,
  and the tokens for A share the same device ID, Gaia can know that B
  and C are in fact on the same device.
To address this problem, the device ID of the source profile is reset
when the token is exported.

Bug: 813928
Change-Id: I92253dcebd0f09b6b4d5a6145e937bea50f5b8a0
Reviewed-on: https://chromium-review.googlesource.com/c/1370275
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#618509}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f9493425cd5b0d4154071f8d5347853acfbe24e