[CrOS MultiDevice] Fix crash when password entered on SL sign-in.
When smart lock is enabled, but the lock screen is disabled on the host,
the smart lock metrics recorder crashes as the user enters their
password.  This is because the case PHONE_NOT_LOCKABLE -- which details
this event -- in the ScreenlockState is not considered in the function
which records these metrics, defaulting in a NOTREACHED().

Bug: 915981
Change-Id: If5a022fcea78e6a46739612e3e101a3eec38a399
Reviewed-on: https://chromium-review.googlesource.com/c/1383201
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#620017}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a31c6b969dd19dbc395c46dacfb9852ce6612e91