Smart Lock: Remove misleading auth failure reason metric bucket.
The state of "Already attempting auth" isn't a true failure;
it means that an attempt was ignored, however, an earlier attempt
is being finished. We only want to record the result of that
earlier attempt, so do not record a failure in this case.

Bug: 905438
Change-Id: Ic8a755c77756e82f997d8dc2f31fa5f98566f8c0
Reviewed-on: https://chromium-review.googlesource.com/c/1385118
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Jeremy Klein <jlklein@chromium.org>
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#621400}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 59bdfdc9aa5cf9148bcd9dcb7cd143111a407443