Fix UseCounter DownloadInSandbox for mainframe download
We were not recording UseCounter DownloadInSandbox for navigation download in
sandboxed mainframe. This CL fixes the case where the download happens in a RFH
that has at least a committed load.

A separate CL will be addressing the case for a popup to a download where there
hasn't been a commit, in which case we might want to / have to record the
UseCounter in the opener page. Currently for this siutation, no UseCounter will
be recorded.

This CL also adds the sandbox bit to the mainframe download UMA/UKM.

Bug: 917082
Change-Id: If61a87f8e12c1e6c756a79da02f2f92f0cb4f5cb
Reviewed-on: https://chromium-review.googlesource.com/c/1383713
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Pavel Feldman <pfeldman@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#621623}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0492d934bfa096c4b26ea03014510d9bc09e15aa