Integrate HintCacheStore into Previews
The HintCache has been updated to use a HintCacheStore, which it owns.
It uses the hint backing store to query for the availability of hints
on the client and to asynchronously load hints on demand. The cache also
contains an MRU memory cache, which retains the most recently loaded
hints for synchronous access.

PreviewsOptimizationGuide creates both the hint cache and
HintCacheLevelDBStore (the concrete derived class of HintCacheStore that
we currently use) on construction and owns the hint cache. This allows
the logic to retain the same cache and store through component updates,
something that would not be possible if PreviewsHints owned them. The
optimization guide does not register for hint component updates until
after the cache is fully initialized, guaranteeing that all information
that PreviewsHints::Create() needs for component processing is
available.

The component processing logic itself in PreviewsHints::Create() has
been modified to use a ComponentUpdateData object provided by the hint
cache store. This will be a nullptr when the component is not newer than
the cache already has, allowing all component hint processing to be
skipped; otherwise, hints from the component are moved into
ComponentUpdateData, allowing us to eliminate superfluous hint copies
and allocations and to minimize processing that must occur on the UI
thread.

The NOSCRIPT logic in
PreviewsDeciderImpl::ShouldAllowPreviewPerOptimizationHints has been
modified to match the RESOURCE_LOADING_HINTS logic to account for the
fact that NoScript page hints may need to be asynchronously loaded. Both
types now rely on the commit time check.

Additionally, a command line switch, "purge_hint_cache_store", has been
added. When it is included, the existing HintCacheStore data is purged
during startup.

Various unittests have been updated to account for the fact that loading
hints from HintCache is now asynchronous and new unittests have been
added to test the new functionality.

Bug: 888097

Change-Id: I258980322b3ae2cfbfd4cc69bd536262ed1049f4
Reviewed-on: https://chromium-review.googlesource.com/c/1402283
Commit-Queue: Jered Gray <jegray@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Doug Arnett <dougarnett@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#623460}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b7c45f89ed5f3fdd3a04fe8f35ae6fc3ecb8947