[Background Fetch] UKM for user initiated aborts.
This adds the logic to log a UKM event, BackgroundFetchDeletingRegistration.

We look for the origin of background fetch in the HistoryService database.
If there's a match, a new SourceId with the origin as the URL is used to log
the UKM data.

UKM collection review:
https://docs.google.com/document/d/1BdKDxQnhanCqzgeflH_V71HuepcZtJwzzDqmgU85JeA/edit?usp=sharing

Bug: 888454
Change-Id: Ie17dd97fc60614e3c7e1517e6dab84b6fe7b461b
Reviewed-on: https://chromium-review.googlesource.com/c/1391916
Commit-Queue: Mugdha Lakhani <nator@google.com>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#624794}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f367101bcdb7a7045d91721965e8bd756b1458e