Add a UserInitiated boolean to PageLoad UKM
This counts a load as user-initiated if it's started in the browser, or
it's started in the renderer and we can tie it back to the original
input event. There are some notable missing cases in the renderer, like
back/forward and reload.

Bug: 921710
Change-Id: Ibc034fbfa077a73148c9cc93032e8dc4a57dae5f
Reviewed-on: https://chromium-review.googlesource.com/c/1423110
Commit-Queue: Annie Sullivan <sullivan@chromium.org>
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#624982}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 496d10a4755317cfeac6349bf88d6270416c37e3