Orderfile: add support for generating a 64 bit orderfile.
Note that the orderfile will be produced in orderfile.arm64.out,
which is currently not used when building (orderfile.arm.out is
used regardless of target_cpu). This makes sense as we use the
same orderfile on all architectures, but is potentially confusing.

Bug: 922460
Change-Id: I5f4b627529d3f9a4f20115a2609e512d83c4b88c
Reviewed-on: https://chromium-review.googlesource.com/c/1417833
Commit-Queue: Matthew Cary <mattcary@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Reviewed-by: Matthew Cary <mattcary@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#625587}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 53f74baf9ad2d42af6c04ec57305955bd24bdfb8