IDL lexer: Update regexp of IDL identifier tokens
WebIDL spec was updated to accept a leading hyphen in identifier
tokens. This CL follows the update.
https://github.com/heycam/webidl/pull/633

Note that this CL makes the lexer accept '-foo' style as a *token*,
so we need more updates to accept it as an *identifier*.


Bug: 850972
Change-Id: Ia484a02380db137ac7a9528ab8b63fd9d0ded152
Reviewed-on: https://chromium-review.googlesource.com/c/1447392
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#627795}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 94b8830b17a394bd090d5592435123b827310461