Fix PasswordManager.FillingAssistance for blacklisted sites
If a user has blacklisted a site for saving credentials, the login database
stores this as an entry with an empty password. This CL recognizes this and
reports that the password was typed while the site was blacklisted instead of
reporting that a password was typed even though a password exits (the empty
password).

There are still two special cases left:

1) If a PSL matching domain is blacklisted, it is still considered part of the
current site even though the user will see a save prompt.

2) The use user has ignored a password save prompt 3 times, we don't ask again.
In this case we will not report the site as blacklisted but report that no
password was stored.

Bug: 918846
Change-Id: I5cb0396940ec6be55b9e0715b3a0a74aa52815b9
Reviewed-on: https://chromium-review.googlesource.com/c/1448213
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Commit-Queue: Dominic Battré <battre@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#627953}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e031efefcef9c1482df55ecac3f485c37f240db9