[Passwords] Introduce PasswordManager.FillingAssistance suffixes
This change introduces a histogram suffixes for
PasswordManager.FillingAssistance that detail whether the form was
filled on a secure or insecure origin. This is useful for experiments
that only target a subset of origins, for example the ongoing HTTP
FOAS experiment.

Bug: 918846, 914765
Change-Id: I6ae96701bbbbd2b364e3e52c4f1ec35e163f499f
Reviewed-on: https://chromium-review.googlesource.com/c/1446155
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#628062}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5a39502e15ac0f3dd880590e423461eb72802d2d