HTTPS Previews: Guard against redirect loops
It is possible for an origin to initiate a redirect loop and for Chrome
to continually follow it, bypassing the normal checks for such a case
because of the navigation restarting of this preview.
This CL protects against that happening.

This is really tricky to test. The given browsertest is the result of a
lot of trial and error where nothing else really works. I also tested
manually with the repro case in the bug.

Bug: 928914
Change-Id: I7cf880afa0581407b8c654bbd4675dbceb78b902
Reviewed-on: https://chromium-review.googlesource.com/c/1455679
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#629636}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f3ca120d90f6a537402935515e69292508b030ae