Define admin policy for reboot on sign out
This is a device policy that specifies whether and in which case the device
must reboot on user sign out.

By default the policy value is NEVER. Two other options are available, with
possibility to add more in the future.

The implementation of the effects of the policy, the reboot itself, will
be done in a separate CL.

Bug: b:77799573
Test: unit_tests --gtest_filter=DeviceSettingsProviderTest.DeviceRebootAfterUserSignout
Change-Id: I2f8e847620f5b28d36d20d3e6912b99ec6e00497
Reviewed-on: https://chromium-review.googlesource.com/c/1455961
Commit-Queue: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Bartosz Fabianowski <bartfab@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#629962}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d2f7ee0b914e71c514b02e37484c25b41bcd5b92