[RELAND] Add metrics for Bluetooth adapter availability
This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/1437721 which
was reverted due to a startup regression on Windows. The only change
in this CL is to disable it on Windows, pending resolution of
https://crbug.com/929375, and also downgrade the priority of sending
metrics to make it a post startup task.

TBR=isherman@chromium.org,rsesek@chromium.org

Bug: 907279
Change-Id: I3c3182901b2f8a753646405699697661c3a12c3b
Reviewed-on: https://chromium-review.googlesource.com/c/1460871
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#630521}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 881a29ab4ec46d558edaea2b21076a6fec551257